Split lockfile regeneration by responsibility
Extract manifest merging, path validation, and Cargo execution into
cohesive modules while retaining `bump_lockfiles` as the stable public
façade and repository boundary.

Record the completed size refactor and final module measurements in the
living ExecPlan.
