+++ lading/commands/bump_lockfile_paths.py
@@ -1,0 +1,51 @@
+"""Path validation and projection for Cargo lockfile regeneration."""
+
+from __future__ import annotations
+
+import collections.abc as cabc
+from pathlib import Path
+
+from lading.exceptions import LadingError
+
+
+class LockfileRegenerationError(LadingError):
+    """Raise when lockfile regeneration cannot validate or execute."""
+
+
+def resolve_lockfile_paths(
+    workspace_root: Path,
+    lockfile_manifests: cabc.Sequence[str],
+) -> tuple[Path, ...]:
+    """Return lockfile paths implied by configured manifest paths."""
+    manifests = _resolve_manifest_paths(workspace_root, lockfile_manifests)
+    return tuple(manifest.parent / "Cargo.lock" for manifest in manifests)
+
+
+def _resolve_manifest_paths(
+    workspace_root: Path,
+    lockfile_manifests: cabc.Sequence[str],
+) -> tuple[Path, ...]:
+    """Return validated root and configured manifest paths in execution order."""
+    resolved_root = workspace_root.resolve()
+    root_manifest = (workspace_root / "Cargo.toml").resolve()
+    seen_manifests: set[Path] = {root_manifest}
+    manifests = [root_manifest]
+    for manifest in lockfile_manifests:
+        candidate = (workspace_root / manifest).resolve()
+        try:
+            candidate.relative_to(resolved_root)
+        except ValueError as exc:
+            message = (
+                f"Lockfile manifest path must stay within the workspace: {manifest}"
+            )
+            raise LockfileRegenerationError(message) from exc
+        if candidate.name != "Cargo.toml":
+            message = (
+                f"Lockfile manifest path must point to a Cargo.toml file: {manifest}"
+            )
+            raise LockfileRegenerationError(message)
+        if candidate in seen_manifests:
+            continue
+        seen_manifests.add(candidate)
+        manifests.append(candidate)
+    return tuple(manifests)
